+++ src/main/java/de/samply/aop/ProjectRoleConstraintsAspect.java
@@ -1,6 +1,7 @@
 package de.samply.aop;
 
 import de.samply.annotations.RoleConstraints;
+import de.samply.annotations.StateConstraints;
 import de.samply.utils.AspectUtils;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -29,14 +30,19 @@ public class ProjectRoleConstraintsAspect {
     @Around("roleConstraintsPointcut()")
     public Object aroundRoleConstraints(ProceedingJoinPoint joinPoint) throws Throwable {
         Optional<RoleConstraints> roleConstraints = fetchRoleConstrains(joinPoint);
+        Optional<StateConstraints> stateConstraints = fetchStateConstrains(joinPoint);
         Optional<String> projectCode = AspectUtils.fetchProjectCode(joinPoint);
         Optional<String> bridgehead = AspectUtils.fetchBridgehead(joinPoint);
-        Optional<ResponseEntity> result = this.constraintsService.checkProjectRoleConstraints(roleConstraints, projectCode, bridgehead);
+        Optional<ResponseEntity> result = this.constraintsService.checkProjectRoleConstraints(roleConstraints, stateConstraints, projectCode, bridgehead);
         return (result.isEmpty()) ? joinPoint.proceed() : result.get();
     }
 
     private Optional<RoleConstraints> fetchRoleConstrains(JoinPoint joinPoint) {
         return AspectUtils.fetchT(joinPoint, RoleConstraints.class);
+    }
+
+    private Optional<StateConstraints> fetchStateConstrains(JoinPoint joinPoint) {
+        return AspectUtils.fetchT(joinPoint, StateConstraints.class);
     }